fix(docs): Update rubric format in README example
When we run the https://github.com/Laboratoria/curriculum-parser#parse-a-project-within-the-curricula-js-repo
example it will error to `Invalid rubric version range`
because the example uses a number that Node won't type coerce to string.
By using a semver like example we will make Node understand that
it should be a string and that `semver.validRange` will get it right.