chore(deps): Replaces cheerio with jsdom
Closes #23

BREAKING CHANGE: HTML fragments processed with jsdom no longer have
their special characters converted to html entities. This is not a
change in API, but strings produced will be different. This should not
"break" web interfaces using the topic JSON, but this may affect search
funcionality in the API/LMS (cc/ @Xpktro).

This affects course/topic readmes in general and quizzes in particular.